+++ dbt/linker.py
@@ -28,6 +28,17 @@ class Linker(object):
     def get_node(self, node):
         return self.graph.node[node]
 
+    def find_cycles(self):
+        try:
+            cycles = nx.algorithms.find_cycle(self.graph)
+        except nx.exception.NetworkXNoCycle:
+            return None
+
+        if cycles:
+            return " --> ".join([".".join(node) for node in cycles])
+
+        return None
+
     def as_topological_ordering(self, limit_to=None):
         try:
             return nx.topological_sort(self.graph, nbunch=limit_to)
@@ -102,11 +113,6 @@ class Linker(object):
         self.graph.add_node(node1)
         self.graph.add_node(node2)
         self.graph.add_edge(node2, node1)
-
-        if len(list(nx.simple_cycles(self.graph))) > 0:
-            raise ValidationException(
-                "Detected a cycle when adding dependency from {} to {}"
-                .format(node1, node2))
 
     def add_node(self, node):
         self.graph.add_node(node)